fastapi_server/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic[email]
python-multipart
python-dotenv
